final/requirements.txt
python-gitlab
requests
jsonschema
pyyaml